refactor: organize all analyzers into vertical slice folders
Move ALIS001-003 from root into domain folders matching the pattern
established by ALIS004 (ControlFlow/) and ALIS005-006 (Validation/):

- ALIS001 → ConditionalChain/ (namespace Alis.Reactive.Analyzers.ConditionalChain)
- ALIS002 → NativeActionLink/ (namespace Alis.Reactive.Analyzers.NativeActionLink)
- ALIS003 → ReactiveEvent/ (namespace Alis.Reactive.Analyzers.ReactiveEvent)

Test files moved to matching folders with updated namespaces.
65 tests pass.
